Fix Issue #360
On Windows, when using newpage, the previewer could randomly stop
working.

On Windows, event 'exit', is not guaranteed to be triggered when a child process
receives a SIGINT, thus causing some promises getting stuck, as they used the handler for
event 'exit' to resolve.

To fix, when the detected platform is Windows, the task killing function
execs "taskkill" to kill the processes, thus avoiding getting stuck
waiting for the 'exit' event handler.